Broke classes, enums, and utilities out into their own library.
