+++ PaladinsApiPCL/PaladinsApiPCL.Core.Domain/Enumerations/QueueType.cs
@@ -1,0 +1,40 @@
+﻿namespace PaladinsApiPCL.Core.Domain.Enumerations
+{
+    public enum QueueType
+    {
+        Challenge_K_StoneKeep = 423,
+        Casual = 424,
+        PracticeVSAI = 425,
+        ChallengeMatch = 426,
+        Practice = 427,
+        SoloCompetitive = 428,
+        zzRETIRED = 429,
+        Challenge_F_TimberMI = 430,
+        Challenge_F_Dock = 431,
+        Challenge_I_Igloo = 432,
+        Challenge_T_Isle = 433,
+        ShootingRange = 434,
+        PerfCaptureMap = 435,
+        TencentAlphaTestQueueCoop = 436,
+        Payload = 437,
+        Challenge_T_Temple = 438,
+        Challenge_I_Mine = 439,
+        Challenge_T_Beach = 440,
+        Challenge_TP = 441,
+        Challenge_FP = 442,
+        Challenge_IP = 443,
+        Tutorial = 444,
+        LiveTestMaps = 445,
+        PvEHandsThatBind = 446,
+        WIPPvELosPollosFernandos = 447,
+        WIPPvEHighRollers = 448,
+        PvEHnS = 449,
+        WIPPvELeapFrogs = 450,
+        PvESurvival = 451,
+        LIVESurvival = 452,
+        LIVESurvivalPractice = 453,
+        Challenge_I_SurvivalRitual = 454,
+        Challenge_T_SurvivalArena = 455,
+        Multi_Queue = 999
+    }
+}
