+++ PaladinsApiPCL/PaladinsApiPCL.Core.Domain/Enumerations/eLanguageCode.cs
@@ -1,0 +1,15 @@
+﻿namespace PaladinsApiPCL.Core.Domain.Enumerations
+{
+    public enum eLanguageCode
+    {
+        English = 1,
+        German = 2,
+        French = 3,
+        Spanish = 7,
+        Spanish_Latin_America = 9,
+        Portuguese = 10,
+        Russian = 11,
+        Polish = 12,
+        Turkish = 13
+    }
+}
