+++ PaladinsApiPCL/PaladinsApiPCL.Core.Domain/Models/Friend.cs
@@ -1,0 +1,9 @@
+﻿namespace PaladinsApiPCL.Core.Domain.Models
+{
+    public class Friend : PaladinsResponse
+    {
+        public string account_id { get; set; }
+        public string name { get; set; }
+        public string player_id { get; set; }
+    }
+}
