+++ PaladinsApiPCL/PaladinsApiPCL.Core.Domain/Models/LoadoutItem.cs
@@ -1,0 +1,9 @@
+﻿namespace PaladinsApiPCL.Core.Domain.Models
+{
+    public class LoadoutItem
+    {
+        public int ItemId { get; set; }
+        public string ItemName { get; set; }
+        public int Points { get; set; }
+    }
+}
