+++ PaladinsApiPCL/PaladinsApiPCL.Core.Domain/Models/MatchDetails.cs
@@ -1,0 +1,66 @@
+﻿namespace PaladinsApiPCL.Core.Domain.Models
+{
+    public class MatchDetails : PaladinsResponse
+    {
+        public int Account_Level { get; set; }
+        public int ActiveId1 { get; set; }
+        public int ActiveId2 { get; set; }
+        public int Assists { get; set; }
+        public string Ban1 { get; set; }
+        public int Ban1Id { get; set; }
+        public string Ban2 { get; set; }
+        public int Ban2Id { get; set; }
+        public string Ban3 { get; set; }
+        public int Ban3Id { get; set; }
+        public string Ban4 { get; set; }
+        public int Ban4Id { get; set; }
+        public int Damage_Bot { get; set; }
+        public int Damage_Done_Magical { get; set; }
+        public int Damage_Done_Physical { get; set; }
+        public int Damage_Mitigated { get; set; }
+        public int Damage_Player { get; set; }
+        public int Damage_Taken { get; set; }
+        public int Deaths { get; set; }
+        public string Entry_Datetime { get; set; }
+        public int Final_Match_Level { get; set; }
+        public int GodId { get; set; }
+        public int Gold_Earned { get; set; }
+        public int Gold_Per_Minute { get; set; }
+        public int Healing { get; set; }
+        public int ItemId1 { get; set; }
+        public int ItemId2 { get; set; }
+        public int ItemId3 { get; set; }
+        public int ItemId4 { get; set; }
+        public int ItemId5 { get; set; }
+        public int ItemId6 { get; set; }
+        public string Item_Active_1 { get; set; }
+        public string Item_Active_2 { get; set; }
+        public string Item_Active_3 { get; set; }
+        public string Item_Purch_1 { get; set; }
+        public string Item_Purch_2 { get; set; }
+        public string Item_Purch_3 { get; set; }
+        public string Item_Purch_4 { get; set; }
+        public string Item_Purch_5 { get; set; }
+        public string Item_Purch_6 { get; set; }
+        public int Killing_Spree { get; set; }
+        public int Kills_Bot { get; set; }
+        public int Kills_Player { get; set; }
+        public int Mastery_Level { get; set; }
+        public int Match { get; set; }
+        public int Minutes { get; set; }
+        public int Multi_kill_Max { get; set; }
+        public int PartyId { get; set; }
+        public string Reference_Name { get; set; }
+        public string Skin { get; set; }
+        public int SkinId { get; set; }
+        public int Structure_Damage { get; set; }
+        public string Surrendered { get; set; }
+        public int Team1Score { get; set; }
+        public int Team2Score { get; set; }
+        public int TeamId { get; set; }
+        public string Team_Name { get; set; }
+        public string Win_Status { get; set; }
+        public string name { get; set; }
+        public string playerName { get; set; }
+    }
+}
