+++ PaladinsApiPCL/PaladinsApiPCL.Core.Domain/Models/MatchId.cs
@@ -1,0 +1,8 @@
+﻿namespace PaladinsApiPCL.Core.Domain.Models
+{
+    public class MatchId : PaladinsResponse
+    {
+        public string Active_Flag { get; set; }
+        public int Match { get; set; }
+    }
+}
