+++ PaladinsApiPCL/PaladinsApiPCL.Core.Domain/Models/PaladinsResponse.cs
@@ -1,0 +1,7 @@
+﻿namespace PaladinsApiPCL.Core.Domain.Models
+{
+    public class PaladinsResponse
+    {
+        public string ret_msg { get; set; }
+    }
+}
