+++ PaladinsApiPCL/PaladinsApiPCL.Core.Domain/Models/PlayerAchievements.cs
@@ -1,0 +1,30 @@
+﻿namespace PaladinsApiPCL.Core.Domain.Models
+{
+    public class PlayerAchievements : PaladinsResponse
+    {
+        public int AssistedKills { get; set; }
+        public int CampsCleared { get; set; }
+        public int DivineSpree { get; set; }
+        public int DoubleKills { get; set; }
+        public int FireGiantKills { get; set; }
+        public int FirstBloods { get; set; }
+        public int GodLikeSpree { get; set; }
+        public int GoldFuryKills { get; set; }
+        public int Id { get; set; }
+        public int ImmortalSpree { get; set; }
+        public int KillingSpree { get; set; }
+        public int MinionKills { get; set; }
+        public string Name { get; set; }
+        public int PentaKills { get; set; }
+        public int PhoenixKills { get; set; }
+        public int PlayerKills { get; set; }
+        public int QuadraKills { get; set; }
+        public int RampageSpree { get; set; }
+        public int ShutdownSpree { get; set; }
+        public int SiegeJuggernautKills { get; set; }
+        public int TowerKills { get; set; }
+        public int TripleKills { get; set; }
+        public int UnstoppableSpree { get; set; }
+        public int WildJuggernautKills { get; set; }
+    }
+}
