+++ PaladinsApiPCL/PaladinsApiPCL.Core.Domain/Models/PlayerLoadouts.cs
@@ -1,0 +1,15 @@
+﻿using System.Collections.Generic;
+
+namespace PaladinsApiPCL.Core.Domain.Models
+{
+    public class PlayerLoadouts : PaladinsResponse
+    {
+        public int ChampionId { get; set; }
+        public string ChampionName { get; set; }
+        public int DeckId { get; set; }
+        public string DeckName { get; set; }
+        public List<LoadoutItem> LoadoutItems { get; set; }
+        public int playerId { get; set; }
+        public string playerName { get; set; }
+    }
+}
