+++ PaladinsApiPCL/PaladinsApiPCL.Core.Domain/Models/Session.cs
@@ -1,0 +1,8 @@
+﻿namespace PaladinsApiPCL.Core.Domain.Models
+{
+    public class Session : PaladinsResponse
+    {
+        public string session_id { get; set; }
+        public string timestamp { get; set; }
+    }
+}
