+++ PaladinsApiPCL/PaladinsApiPCL.Core.Domain/PaladinsApiPCL.Core.Domain.csproj
@@ -1,0 +1,7 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netcoreapp1.1</TargetFramework>
+  </PropertyGroup>
+
+</Project>
