+++ PaladinsApiPCL/PaladinsApiPCL.Core.sln
@@ -3,9 +3,11 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
 VisualStudioVersion = 15.0.26430.6
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PaladinsApiPCL-Core", "PaladinsApiPCL-Core\PaladinsApiPCL-Core.csproj", "{024E2256-C123-4998-B5E1-D382484EE531}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PaladinsApiPCL.Core", "PaladinsApiPCL.Core\PaladinsApiPCL.Core.csproj", "{024E2256-C123-4998-B5E1-D382484EE531}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PaladinsApiPCL-Core.Tests", "PaladinsApiPCL-Core.Tests\PaladinsApiPCL-Core.Tests.csproj", "{375B4F6C-67A3-4202-ADB0-4FF72A6F6A08}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PaladinsApiPCL.Core.Tests", "PaladinsApiPCL.Core.Tests\PaladinsApiPCL.Core.Tests.csproj", "{375B4F6C-67A3-4202-ADB0-4FF72A6F6A08}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PaladinsApiPCL.Core.Domain", "PaladinsApiPCL.Core.Domain\PaladinsApiPCL.Core.Domain.csproj", "{71F08F32-F3A2-48A0-B18B-3F014E115AA0}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -21,6 +23,10 @@ Global
 		{375B4F6C-67A3-4202-ADB0-4FF72A6F6A08}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{375B4F6C-67A3-4202-ADB0-4FF72A6F6A08}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{375B4F6C-67A3-4202-ADB0-4FF72A6F6A08}.Release|Any CPU.Build.0 = Release|Any CPU
+		{71F08F32-F3A2-48A0-B18B-3F014E115AA0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{71F08F32-F3A2-48A0-B18B-3F014E115AA0}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{71F08F32-F3A2-48A0-B18B-3F014E115AA0}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{71F08F32-F3A2-48A0-B18B-3F014E115AA0}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
