+++ PaladinsApiPCL/PaladinsApiPCL.Core/PaladinsApiPCL.Core.csproj
@@ -8,4 +8,8 @@
     <PackageReference Include="Newtonsoft.Json" Version="10.0.2" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\PaladinsApiPCL.Core.Domain\PaladinsApiPCL.Core.Domain.csproj" />
+  </ItemGroup>
+
 </Project>
